@@ -1,11 +1,5 @@
 module github.com/fsgo/fspool
 
-go 1.20
+go 1.21
 
-require github.com/stretchr/testify v1.8.4
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
+require github.com/fsgo/fst v0.0.3
