@@ -8,33 +8,33 @@ import (
 	"net"
 	"testing"
 
-	"github.com/stretchr/testify/require"
+	"github.com/fsgo/fst"
 )
 
 func TestReadMeta(t *testing.T) {
 	t.Run("conn", func(t *testing.T) {
 		pc := newPConn(nil, nil)
 		m := ReadMeta(pc)
-		require.NotNil(t, m)
+		fst.NotNil(t, m)
 	})
 
 	t.Run("testMeta1 has", func(t *testing.T) {
 		pc := newPConn(nil, nil)
 		n1 := &testMeta1{Conn: pc}
 		m := ReadMeta(n1)
-		require.NotNil(t, m)
+		fst.NotNil(t, m)
 	})
 
 	t.Run("nil", func(t *testing.T) {
 		rc := &net.TCPConn{}
 		m := ReadMeta(rc)
-		require.Nil(t, m)
+		fst.Nil(t, m)
 	})
 
 	t.Run("testMeta1 nil", func(t *testing.T) {
 		n1 := &testMeta1{Conn: nil}
 		m := ReadMeta(n1)
-		require.Nil(t, m)
+		fst.Nil(t, m)
 	})
 }
 
@@ -44,6 +44,6 @@ type testMeta1 struct {
 	net.Conn
 }
 
-func (tm *testMeta1) PERaw() interface{} {
+func (tm *testMeta1) PERaw() any {
 	return tm.Conn
 }
